+++ Tests/NPlusOneQueryTests/NPlusOneQueryVerifier.cs
@@ -1,0 +1,11 @@
+using Collections.Analyzer.Diagnostics.CI0011;
+using Microsoft.CodeAnalysis.Testing;
+
+namespace Tests.NPlusOneQueryTests;
+
+public class NPlusOneQueryVerifier : AnalyzerVerifier<
+    NPlusOneQueryAnalyzer,
+    NPlusOneQueryTests,
+    DefaultVerifier>
+{
+}