+++ Tests/Resources/NPlusOneQuery/NPlusOneQuery1.cs
@@ -1,0 +1,19 @@
+using System;
+using System.Collections.Generic;
+
+public class TestClass
+{
+    public void TestMethod(IEnumerable<int> items)
+    {
+        var repository = new MyRepository();
+        foreach (var item in items)
+        {
+            var data = {|CI0011:repository.GetData(item)|};
+        }
+    }
+
+    public class MyRepository
+    {
+        public string GetData(int id) => id.ToString();
+    }
+}